requirements.txt
PyJWT==1.5.2
Flask==0.12.2
Flask-Cors==3.0.3
requests==2.18.1
requests-oauthlib==0.8.0
